requirements.txt
# Core frameworks
tensorflow==2.15.0
flask==3.0.0
flask-cors==4.0.0

# Data processing
numpy==1.24.0
pandas==2.2.2
opencv-python==4.9.0.80
Pillow==10.2.0
scipy==1.11.3

# Utility tools
gunicorn==21.2.0
python-dotenv==1.0.1

# Optional: For model saving/loading
h5py==3.10.0
